win_file.h: replace parameter type to uint about Read and Write File.

+++ include/win_file.h
@@ -8,10 +8,10 @@
 
 // The buffer allocated from ReadFile must call Runtime_M.Memory.Free().
 
-typedef errno (*WFReadFileA_t)(LPSTR path, byte** buf, int64* size);
-typedef errno (*WFReadFileW_t)(LPWSTR path, byte** buf, int64* size);
-typedef errno (*WFWriteFileA_t)(LPSTR path, byte* buf, int64 size);
-typedef errno (*WFWriteFileW_t)(LPWSTR path, byte* buf, int64 size);
+typedef errno (*WFReadFileA_t)(LPSTR path, byte** buf, uint* size);
+typedef errno (*WFReadFileW_t)(LPWSTR path, byte** buf, uint* size);
+typedef errno (*WFWriteFileA_t)(LPSTR path, byte* buf, uint size);
+typedef errno (*WFWriteFileW_t)(LPWSTR path, byte* buf, uint size);
 
 typedef errno (*WFUninstall_t)();
